Got MAC and WWPN pools working
Leaving these a bit statically defined now for brevity. I will be circling back to make thing a little more dynamic later. For instance, I'm only creating one big MAC pool for now, and will be mapping all vNIC templates to it. Later it may be useful to create multiple MAC pools for different use cases, like I usually do when I'm configuring UCS manually. However, it's a bit more complexity than it's worth for now, so I'm going to circle back and re-evaluate this later to see if it's worth it.