Got createBootPolicy working
each element could get it's own try block, but for now I'm leaving it all inside one big block. So if there's a change in WWPNs to the same boot policy, that's not going to work right now. I'm going to circle back and fix this, but only after I think of a way to simplify the boot policy creation process, possibly by integrating with netapp in some way.