Added vNIC and vHBA Template methods.
Consider the mechanism by which these are created a "work in progress". Right now it's based off of a statically defined dictionary. This isn't a big deal, since virtual NICs can be plentifully created in UCS, but still there might need to be a bit more flexibility from a user experience perspective. So this may be revisited.