Some minor changes to Nexus templates.
Decided to chunk these templates up __a little__ in the future. There will be one large template for all of the config that's not really that dynamic. Pretty much everything pulled from the YAML file. However, I will pull a lot from other systems as well, such as the WWPNs from Netapp and UCS. More on this later.