First "go" at new structure for UCS functions
I am building a new home for UCS functions designed to function similarly to the PowerShell equivalents. I also built a prototype for deleting objects from the configuration. With this code, I can now remove a VLAN or VSAN from the YAML file, and it will be subsequently deleted from UCS - so it's not just a one-time configuration.

For now this is limited to the VLANs and VSANs, and I commented everything else out for now. I will be building the rest of the functions the same way